+++ src/main/java/tensor/DataType.java
@@ -3,7 +3,7 @@ package tensor;
 public enum DataType {
     FLOAT64, // double v Javě
     FLOAT32, // float v Javě
-    FLOAT16, // v Javě se často reprezentuje jako short[]
+    BFLOAT16, // v Javě se často reprezentuje jako short[]
     INT32,   // integer index tensor stored explicitly
     BOOL     // mask/boolean tensor stored explicitly
 }
